Fix version display in commands and log messages
Previously it tried to match the next lowest pre-defined version and
displayed that in the command/log output as well. The plugin doesn't
use the selection logic like that anymore when deciding which server
to connect to so the messages should always output the real protocol
version number and only a pre-defined constant if it was found.